Revert "Added some tests"
This reverts commit 5a7b093e3cc47494b6117c6613091500296a79b2.

+++ src/common/assignment.cpp
@@ -5,7 +5,7 @@
 
 using namespace gd;
 
-VertexAssignment::VertexAssignment(size_t num_vertices)
+VertexAssignment::VertexAssignment(int num_vertices)
 {
   m_assignment.resize(num_vertices);
   std::fill(m_assignment.begin(), m_assignment.end(), UINT_UNDEF);
