+++ src/common/assignment.hpp
@@ -9,7 +9,7 @@ namespace gd
   class VertexAssignment
   {
     public:
-      VertexAssignment(size_t num_vertices);
+      VertexAssignment(int num_vertices);
 
       void assign(vertex_t vertex, point_id_t to);
       bool isAssigned(vertex_t vertex) const;
